+++ ratelimit/ratelimit_test.go
@@ -17,7 +17,7 @@ func TestRateLimit(t *testing.T) {
 		t.Log("Note: The TestRateLimit test is known to have a bug if run near the top of the hour. Since the rate limiter isn't a moving window, it could end up checking against two different buckets on either side of the top of the hour, so if you see that just re-run it after you've passed the top of the hour.")
 	}
 
-	rateLimiter := NewRateLimiter(os.Getenv("REDIS_URL"), fmt.Sprintf("worker-test-rl-%d", os.Getpid()))
+	rateLimiter := NewRateLimiter(os.Getenv("REDIS_URL"), fmt.Sprintf("worker-test-rl-%d", os.Getpid()), false, time.Minute)
 
 	ok, err := rateLimiter.RateLimit(context.TODO(), "slow", 2, time.Hour)
 	if err != nil {